XX-7346 - Remove /var/sipxdata/phonebook references, since it is no longer used
- remove the appearence of /var/sipxdata/phonebook directory in sipx-backup.in so that the back-up operation won't use it anymore.
- modified PhonebookFileEntryTrigger (upgrade task) to delete the /var/sipxdata/phonebook directory after the phonebook entries are saved into db.

Patch from Tambrea Cosmin, thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17726 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/phonebook/PhonebookFileEntryTrigger.java
@@ -34,26 +34,38 @@ public class PhonebookFileEntryTrigger extends InitTaskListener {
         LOG.info("Begining saving entries from existing CSV/vCard files into the table phonebook_file_entry.");
 
         String directory = m_phonebookManager.getExternalUsersDirectory();
+        File dir = new File(directory);
 
-        Map<Integer, String[]> files = m_phonebookManager.getPhonebookFilesName();
+        if (dir.exists()) {
+            Map<Integer, String[]> files = m_phonebookManager.getPhonebookFilesName();
 
-        for (Map.Entry<Integer, String[]> entry : files.entrySet()) {
-            Integer phonebookId = entry.getKey();
-            for (String fileName : entry.getValue()) {
-                if (StringUtils.isBlank(fileName)) {
-                    continue;
-                }
-                try {
-                    File f = new File(directory, fileName);
-                    InputStream fileStream = new FileInputStream(f);
-                    m_phonebookManager.addEntriesFromFile(phonebookId, fileStream);
-                    f.delete();
-                } catch (FileNotFoundException e) {
-                    LOG.warn("File not found error :" + e.getMessage());
+            for (Map.Entry<Integer, String[]> entry : files.entrySet()) {
+                Integer phonebookId = entry.getKey();
+                for (String fileName : entry.getValue()) {
+                    if (StringUtils.isBlank(fileName)) {
+                        continue;
+                    }
+                    try {
+                        File f = new File(directory, fileName);
+                        InputStream fileStream = new FileInputStream(f);
+                        m_phonebookManager.addEntriesFromFile(phonebookId, fileStream);
+                        f.delete();
+                    } catch (FileNotFoundException e) {
+                        LOG.warn("File not found error :" + e.getMessage());
+                    }
                 }
             }
-        }
 
-        m_phonebookManager.removeTableColumns();
+            m_phonebookManager.removeTableColumns();
+
+            if (dir.delete()) {
+                LOG.info("Directory " + directory + "was successfully deleted !");
+            } else {
+                LOG.info("Failed to delete directory " + directory);
+            }
+
+        } else {
+            LOG.info("Phonebook's directory " + directory + " doesn't exist!");
+        }
     }
 }
